Filter some sends

+++ src/nodemon/Services/TelemetrySingleton.cs
@@ -2,7 +2,7 @@
 
 public class TelemetrySingleton
 {
-    public double? ChassisTemperature { get; set; }
-    public double? ChassisHumidity { get; set; }
+    public int? ChassisTemperature { get; set; }
+    public int? ChassisHumidity { get; set; }
     public Dictionary<string, double> PaTemperatures { get; private set; } = [];
 }
